requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
contourpy==1.3.1
cycler==0.12.1
dateutils==0.6.12
deprecation==2.1.0
everapi==0.1.1
Flask==3.1.0
fonttools==4.55.3
freecurrencyapi==0.1.0
frozendict==2.4.6
functions-framework==3.8.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
multitasking==0.0.11
nselib==1.5
nsetools==1.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
scipy==1.15.1
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
yfinance==0.2.52
